refactor(progressions): extract singleMoveDiff to its own module (fix react-refresh lint)
