+++ src/components/SongControls/ProgressionStepList.tsx
@@ -8,6 +8,7 @@ import {
   type ResolvedProgressionStep,
 } from "../../progressions/progressionDomain";
 import { useTranslation } from "../../hooks/useTranslation";
+import { singleMoveDiff } from "./progressionStepListUtils";
 import styles from "./ProgressionStepList.module.css";
 
 interface ProgressionStepListProps {
@@ -22,23 +23,6 @@ interface ProgressionStepListProps {
   caption: string;
   /** Right-aligned summary in the caption row (e.g. "9 chords · 10 bars"). */
   meta?: string;
-}
-
-/**
- * Collapse a full reordered id array (as handed back by motion's Reorder.Group)
- * into the single `{ from, to }` move it represents. Returns null when the arrays
- * are identical or differ by more than one contiguous move.
- */
-export function singleMoveDiff(prev: string[], next: string[]): { from: number; to: number } | null {
-  if (prev.length !== next.length) return null;
-  let lo = 0;
-  while (lo < prev.length && prev[lo] === next[lo]) lo++;
-  let hi = prev.length - 1;
-  while (hi >= 0 && prev[hi] === next[hi]) hi--;
-  if (lo > hi) return null;
-  if (next[hi] === prev[lo]) return { from: lo, to: hi };
-  if (next[lo] === prev[hi]) return { from: hi, to: lo };
-  return null;
 }
 
 interface StepRowProps {
